Treat AND and OR case insensitively... AND == and == And
The purist in my wants to enforce all caps, but that makes the language
less forgiving. It's not the sort of thing someone wants to invest a lot
of time learning, so make it forgiving. If this was a specialty tool for
someone to use with official training, and the expressions were saved
and compared later, then a stricter syntax would make for more standard
expressions. But this is neither for trained users nor for long-lived
expressions, so may as well make it forgiving.